requirements.txt
Django~=3.2.6
argon2-cffi
django-taggit-1.5.1